requirements.txt
# Core dependencies
torch>=1.12.0
mediapipe>=0.9.0
numpy>=1.22.0
opencv-python>=4.5.5
pillow>=9.0.0

# Data processing and visualization
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.0.0
pandas>=1.4.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0